requirements.txt
beautifulsoup4==4.12.2
requests>=2.32.2,<3.0
pydantic==2.5.2
lxml==4.9.3
urllib3>=2.2.2,<3.0
python-dateutil==2.8.2
typing-extensions>=4.8.0
pytest==8.0.2
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1